modifier fraisBancaires data type to double

+++ src/Model/CompteCourant.java
@@ -1,18 +1,18 @@
 package Model;
 
 public class CompteCourant extends Compte {
-    private  int FraisBancaires;
+    private  double FraisBancaires;
 
-    public CompteCourant(int numero, double solde, Client proprietaire, int fraisBancaires) {
+    public CompteCourant(int numero, double solde, Client proprietaire, double fraisBancaires) {
         super(numero, solde, proprietaire);
         FraisBancaires = fraisBancaires;
     }
 
-    public int getFraisBancaires() {
+    public double getFraisBancaires() {
         return FraisBancaires;
     }
 
-    public void setFraisBancaires(int fraisBancaires) {
+    public void setFraisBancaires(double fraisBancaires) {
         FraisBancaires = fraisBancaires;
     }
 }
